Re-throw the thrown error, if possible
