+++ lib/temporal/errors.rb
@@ -26,19 +26,10 @@ module Temporal
 
   class NotFoundFailure < ApiError; end
 
-  # Superclass for errors raised when retrieving a workflow result on the
+  # Superclass for system errors raised when retrieving a workflow result on the
   # client, but the workflow failed remotely.
   class WorkflowError < Error; end
 
-  # The workflow failed in user code.
-  class WorkflowFailed < WorkflowError
-    attr_reader :stack_trace
-
-    def initialize(message, stack_trace:)
-      super(message)
-      @stack_trace = stack_trace
-    end
-  end
   class WorkflowTimedOut < WorkflowError; end
   class WorkflowTerminated < WorkflowError; end
   class WorkflowCanceled < WorkflowError; end
