+++ lib/temporal/workflow/state_manager.rb
@@ -5,6 +5,7 @@ require 'temporal/workflow/command_state_machine'
 require 'temporal/workflow/history/event_target'
 require 'temporal/metadata'
 require 'temporal/concerns/payloads'
+require 'temporal/workflow/errors'
 
 module Temporal
   class Workflow
@@ -145,11 +146,11 @@ module Temporal
 
         when 'ACTIVITY_TASK_FAILED'
           state_machine.fail
-          dispatch(target, 'failed', parse_failure(event.attributes.failure, ActivityException))
+          dispatch(target, 'failed', Temporal::Workflow::Errors.new.error_from(event.attributes.failure, ActivityException))
 
         when 'ACTIVITY_TASK_TIMED_OUT'
           state_machine.time_out
-          dispatch(target, 'failed', parse_failure(event.attributes.failure))
+          dispatch(target, 'failed', Temporal::Workflow::Errors.new.error_from(event.attributes.failure))
 
         when 'ACTIVITY_TASK_CANCEL_REQUESTED'
           state_machine.requested
@@ -161,7 +162,7 @@ module Temporal
 
         when 'ACTIVITY_TASK_CANCELED'
           state_machine.cancel
-          dispatch(target, 'failed', parse_failure(event.attributes.failure))
+          dispatch(target, 'failed', Temporal::Workflow::Errors.new.error_from(event.attributes.failure))
 
         when 'TIMER_STARTED'
           state_machine.start
@@ -224,15 +225,15 @@ module Temporal
 
         when 'CHILD_WORKFLOW_EXECUTION_FAILED'
           state_machine.fail
-          dispatch(target, 'failed', parse_failure(event.attributes.failure))
+          dispatch(target, 'failed', Temporal::Workflow::Errors.new.error_from(event.attributes.failure))
 
         when 'CHILD_WORKFLOW_EXECUTION_CANCELED'
           state_machine.cancel
-          dispatch(target, 'failed', parse_failure(event.attributes.failure))
+          dispatch(target, 'failed', Temporal::Workflow::Errors.new.error_from(event.attributes.failure))
 
         when 'CHILD_WORKFLOW_EXECUTION_TIMED_OUT'
           state_machine.time_out
-          dispatch(target, 'failed', parse_failure(event.attributes.failure))
+          dispatch(target, 'failed', Temporal::Workflow::Errors.new.error_from(event.attributes.failure))
 
         when 'CHILD_WORKFLOW_EXECUTION_TERMINATED'
           # todo
@@ -317,32 +318,6 @@ module Temporal
         end
       end
 
-      def parse_failure(failure, default_exception_class = StandardError)
-        case failure.failure_info
-        when :application_failure_info
-          exception_class = safe_constantize(failure.application_failure_info.type)
-          exception_class ||= default_exception_class
-          message = from_details_payloads(failure.application_failure_info.details)
-          backtrace = failure.stack_trace.split("\n")
-
-          exception_class.new(message).tap do |exception|
-            exception.set_backtrace(backtrace) if !backtrace.empty?
-          end
-        when :timeout_failure_info
-          TimeoutError.new("Timeout type: #{failure.timeout_failure_info.timeout_type.to_s}")
-        when :canceled_failure_info
-          # TODO: Distinguish between different entity cancellations
-          StandardError.new(from_payloads(failure.canceled_failure_info.details))
-        else
-          StandardError.new(failure.message)
-        end
-      end
-
-      def safe_constantize(const)
-        Object.const_get(const) if Object.const_defined?(const)
-      rescue NameError
-        nil
-      end
     end
   end
 end
